% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot fatality case points}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric number identifying the state}

\item{year}{A numeric number repsresenting the year (e.g. 2013)}
}
\value{
A map of the state and plots of fatality case points
}
\description{
This function plots fatality case (accident) points for the state
(using \code{state.num}) and the year (using \code{year}).
}
\note{
An error occurs if the state number is invalid.
Nothing is plotted if no fatality cases exist.
}
\examples{
\dontrun{
fars_map_state(1, 2013)
}

}
