% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read Fatality Analysis Reporting System data by the file name}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string of the name of a file to read.}
}
\value{
a data.frame of the data that is read in.
}
\description{
This function reads a csv file of Fatality Analysis Reporting
System data (using the \code{filename} argument) and returns
a data.frame of the data.
}
\note{
Error occurs if the spcified file does not exist.
}
\examples{
\dontrun{
librray(readr)
library(dplyr)
data <- fars_read("accident_2013.csv.bz2")
}

}
