% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Fatality Analysis Reporting System data for the years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years}
}
\value{
A list of data.frames corresponding to the years
}
\description{
This function reads csv files of Fatality Analysis Reporting
System data for the years (using the \code{years} argument)
and returns a list of data.frames corresponding to the years.
}
\note{
Any invalid year in the vector of years results in Null
item of the returned list with warning messages "invalid year: yyyy"
}
\examples{
\dontrun{
 dat_list <- fars_read_years(c(2013, 2014, 2015))
 }

}
