% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make file name from year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric number repsresenting the year (e.g. 2013)}
}
\value{
A character string of the file name in the form of
"accident_yyyy.csv.bz2"
}
\description{
This function creates a file name of the specified year's
Read Fatality Analysis Reporting System data.
}
\note{
Validity of the year is not checked
}
\examples{
\dontrun{
file <- make_filename(2013)
}

}
